Update DataAccess.Users.cs
Minor fix for an exception being thrown in the DataAccess.Users file.  Currently when we do a lookup of the SudoLogin dictionary it is wrapped in a try {} catch{} that just throws it away the exception.   I added an if statement which reduces how often the exception is being thrown.