Update DataAccess.Utilities.cs
Minor fix for an exception being thrown in the DataAccess.Utilities file.  Currently when we create the defaultValue using the Activator.CreateInstance it is wrapped in a try {} catch{} that just throws it away the exception.   I added an if statement which reduces how often the exception is being thrown when the default type is a string.   

I was noticing that this exception was only being thrown for strings so I added a default value for string of string.Empty and the exception stopped being thrown.  This probably isn't going to cover every case so I left in the try { } catch {}.